DLL moi(cu) (1)/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
mlxtend
shap
papermill
pyyaml
pyarrow
ipykernel
jupyter
streamlit
